Fixed navigation and seperated it into a Header component. Made pages components for each page. Seperated API call logic into reusable component ToDoContainer.
