+++ src/components/ToDoList.module.css
@@ -20,6 +20,7 @@
     padding: 5px;
     margin: 0px;
     height: 3em;
+    justify-content: right;
 }
 
 .Container {
@@ -52,6 +53,7 @@
     width: 20px;
     height: 20px;
     margin: 0px 5px 0px 5px;
+    align-self: flex-end;
 }
 
 body {
