+++ src/components/ToDoListItem.js
@@ -3,9 +3,11 @@ import propTypes from 'prop-types';
 import style from './ToDoList.module.css';
 function ToDoListItem({todo, onRemoveToDo}) {
   //unsure if I need to keep in id={todo.id} in li? works without it
+  console.log(todo);
   return (
     <>
       <li id= {todo.id} className= {style.ListItem}>
+       
       {todo.fields.Title}
       <img src="https://cdn.discordapp.com/attachments/955553213988360212/989219882530922506/trashIcon.png" className= {style.RemoveButton} onClick={() => onRemoveToDo(todo.id)}/>
       </li>
